requirements.txt
asgiref==3.8.1
attrs==25.1.0
Django==5.1.6
django-filter==24.3
django-seed==0.3.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-spectacular==0.28.0
Faker==35.2.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.36.2
rest-framework-roles==1.0.6
rpds-py==0.22.3
six==1.17.0
sqlparse==0.5.3
toposort==1.10
typing_extensions==4.12.2
uritemplate==4.1.1
